Performance: Use Span.CopyTo for 1.0x factor in DirectImageTransfer
Optimized `DirectImageTransfer` by introducing a fast path for identity transfers (factor == 1.0f).
Instead of calculating and copying pixels via nested loops, it now uses `srcSpan.CopyTo(destPixels)`.
This significantly improves memory copy speeds for unscaled images.
